+++ AzeriteUI/Components/ActionBars/Prototypes/ActionBar.lua
@@ -242,8 +242,8 @@ ActionBar.CreateButton = function(self, buttonConfig)
 	button:SetAttribute("statehidden", nil)
 	button:UpdateAction()
 
-	-- Add in a vehicle exit button at slot 7 for Wrath and Retail.
-	if ((ns.IsWrath or ns.IsRetail) and (self.id == 1 and button.id == 7)) then
+	-- Add in a vehicle exit button at slot 7 for Wrath and Retail. CATA: check
+	if ((ns.IsWrath or ns.IsCata or ns.IsRetail) and (self.id == 1 and button.id == 7)) then
 		button:SetState(16, "custom", exitButton)
 		button:SetState(17, "custom", exitButton)
 		button:SetState(18, "custom", exitButton)
@@ -390,6 +390,7 @@ end
 ActionBar.UpdateStateDriver = function(self)
 	if (InCombatLockdown()) then return end
 
+	-- CATA: check stances and states
 	local statedriver
 	if (self.id == 1) then
 		statedriver = "[overridebar] possess; [possessbar] possess; [shapeshift] possess; [bonusbar:5] dragon; [form,noform] 0; [bar:2] 2; [bar:3] 3; [bar:4] 4; [bar:5] 5; [bar:6] 6"
@@ -401,17 +402,17 @@ ActionBar.UpdateStateDriver = function(self)
 			statedriver = statedriver .. "; [bonusbar:1] 7; [bonusbar:2] 8; [bonusbar:3] 9"
 
 		elseif (playerClass == "ROGUE") then
-			if (ns.IsWrath) then
+			if (ns.IsWrath or ns.IsCata) then
 				statedriver = statedriver .. "; [bonusbar:1] 7 [bonusbar:2] 8" -- Shadowdance
 			else
 				statedriver = statedriver .. "; [bonusbar:1] 7"
 			end
 
-		elseif (ns.IsWrath and playerClass == "PRIEST") then
+		elseif ((ns.IsWrath or ns.IsCata) and playerClass == "PRIEST") then
 			statedriver = statedriver .. "; [bonusbar:1] 7" -- Shadowform
 
 		elseif (playerClass == "WARRIOR") then
-			if (ns.IsWrath) then
+			if (ns.IsWrath or ns.IsCata) then
 				statedriver = statedriver .. "; [bonusbar:1] 7; [bonusbar:2] 8; [bonusbar:3] 9"
 			elseif (ns.IsClassic) then
 				statedriver = statedriver .. "; [bonusbar:1] 7; [bonusbar:2] 8"