+++ AzeriteUI/Components/Misc/Minimap.lua
@@ -85,8 +85,8 @@ MinimapMod.GenerateDefaults = function(self)
 	defaults.profile.savedPosition = {
 		scale = mapScale * ns.API.GetEffectiveScale(),
 		[1] = "BOTTOMRIGHT",
-		[2] = -(40 - ((ns.IsWrath or ns.IsClassic) and 10 or 0)) * (mapScale * ns.API.GetEffectiveScale()),
-		[3] = (40 - ((ns.IsWrath or ns.IsClassic) and 10 or 0)) * (mapScale * ns.API.GetEffectiveScale())
+		[2] = -(40 - ((ns.IsCata or ns.IsWrath or ns.IsClassic) and 10 or 0)) * (mapScale * ns.API.GetEffectiveScale()),
+		[3] = (40 - ((ns.IsCata or ns.IsWrath or ns.IsClassic) and 10 or 0)) * (mapScale * ns.API.GetEffectiveScale())
 	}
 	return defaults
 end
@@ -130,18 +130,18 @@ local ObjectSnippets = {
 	},
 	Eye = {
 		Enable = function(object)
-			if (ns.IsWrath) then
+			if (ns.IsWrath or ns.IsCata) then
 				object:SetFrameLevel(object:GetParent():GetFrameLevel() + 2)
 			elseif (ns.IsRetail) then
 			end
 		end,
 		Disable = function(object)
-			if (ns.IsWrath) then
+			if (ns.IsWrath or ns.IsCata) then
 			elseif (ns.IsRetail) then
 			end
 		end,
 		Update = function(object)
-			if (ns.IsWrath) then
+			if (ns.IsWrath or ns.IsCata) then
 			elseif (ns.IsRetail) then
 			end
 		end
@@ -169,7 +169,7 @@ local ObjectSnippets = {
 	------------------------------------------
 	AzeriteEye = {
 		Enable = function(object)
-			if (ns.IsWrath) then
+			if (ns.IsWrath or ns.IsCata) then
 				MiniMapLFGFrame:SetParent(Minimap)
 				MiniMapLFGFrame:SetFrameLevel(100)
 				MiniMapLFGFrame:ClearAllPoints()
@@ -188,7 +188,7 @@ local ObjectSnippets = {
 			end
 		end,
 		Disable = function(object)
-			if (ns.IsWrath) then
+			if (ns.IsWrath or ns.IsCata) then
 				MiniMapLFGFrame:SetParent(_G[ObjectOwners.Eye])
 				MiniMapLFGFrame:SetFrameLevel(MinimapBackdrop:GetFrameLevel() + 2)
 				MiniMapLFGFrame:ClearAllPoints()
@@ -310,7 +310,8 @@ local Skins = {
 				Point = { "CENTER", 0, 0 },
 				Color = { .90, .95, 1 }
 			},
-			AzeriteEyeClassicPvP = (ns.IsClassic or ns.IsWrath) and {
+			-- CATA: check
+			AzeriteEyeClassicPvP = (ns.IsClassic or ns.IsWrath or ns.IsCata) and {
 				Owner = "EyeClassicPvP",
 				DrawLayer = "BORDER",
 				DrawLevel = 2,
@@ -993,7 +994,8 @@ MinimapMod.InitializeObjectTables = function(self)
 		Objects.ZoomOut = Minimap.ZoomOut
 	end
 
-	if (ns.IsWrath) then
+	-- CATA: check
+	if (ns.IsWrath or ns.IsCata) then
 		Objects.BorderTop = MinimapBorderTop
 		Objects.BorderClassic = MinimapBorder
 		Objects.Calendar = GameTimeFrame
@@ -1047,7 +1049,8 @@ MinimapMod.InitializeObjectTables = function(self)
 		ObjectOwners.ZoomOut = Minimap
 	end
 
-	if (ns.IsWrath) then
+	-- CATA: check
+	if (ns.IsWrath or ns.IsCata) then
 		ObjectOwners.BorderTop = MinimapCluster
 		ObjectOwners.BorderClassic = MinimapBackdrop
 		ObjectOwners.Calendar = MinimapCluster