+++ AzeriteUI/Components/Misc/VehicleSeat.lua
@@ -25,7 +25,7 @@
 --]]
 local _, ns = ...
 
-if (not ns.IsRetail and not ns.IsWrath) then return end
+if (not ns.IsRetail and not ns.IsWrath and not ns.IsCata) then return end
 
 local L = LibStub("AceLocale-3.0"):GetLocale((...))