+++ AzeriteUI/Core/API/Colors.lua
@@ -95,7 +95,8 @@ local GetDifficultyColor = function(level, isScaling)
 	end
 end
 
-if (ns.IsClassic or ns.IsTBC or ns.IsWrath) then
+-- CATA: check this
+if (ns.IsClassic or ns.IsTBC or ns.IsWrath or ns.IsCata) then
 	GetDifficultyColor = function(level, isScaling)
 		local colors = Colors.quest
 		local levelDiff = level - UnitLevel("player")