+++ AzeriteUI/Options/OptionsPages/ExplorerMode.lua
@@ -268,14 +268,14 @@ local GenerateOptions = function()
 				hidden = isdisabled,
 				order = 46
 			},
-			fadeWithFocusTarget = (ns.IsRetail or ns.IsWrath) and {
+			fadeWithFocusTarget = (ns.IsRetail or ns.IsWrath or ns.IsCata) and {
 				name = L["While having a focus target"],
 				desc = "",
 				type = "toggle", width = "full",
 				set = setterReverse, get = getterReverse, disabled = isdisabled, hidden = isdisabled,
 				order = 50
 			} or nil,
-			fadeInVehicles = (ns.IsRetail or ns.IsWrath) and {
+			fadeInVehicles = (ns.IsRetail or ns.IsWrath or ns.IsCata) and {
 				name = L["While having any sort of replacement actionbar"],
 				desc = L["Exits Explorer Mode when in vehicles, when possessing a unit or any other situation that would replace your character's action bar with a temporary one."],
 				type = "toggle", width = "full",
@@ -335,7 +335,7 @@ local GenerateOptions = function()
 				set = setter, get = getter, disabled = isdisabled, hidden = isdisabled,
 				order = 141
 			},
-			fadeFocusFrame = (ns.IsRetail or ns.IsWrath) and {
+			fadeFocusFrame = (ns.IsRetail or ns.IsWrath or ns.IsCata) and {
 				name = L["Fade Focus unit frame"],
 				desc = "",
 				type = "toggle", width = "full",