+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file. Be aware th
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] 2024-04-30
+- Cataclysm Classic Pre-Patch.
+
 ## [5.2.169-Release] 2024-04-29
 ### Added
 - Added an option to the chat settings in `/azerite` to toggle whether or not the primary chat window is cleared on logon and reloads. Also added a slider to adjust the number of seconds the window is kept clear.